# src/models/userProject-skill/dto/userProject-skill.update.dto.ts
import { Expose } from 'class-transformer';
import { ApiProperty } from '@nestjs/swagger';
import { IsOptional } from 'class-validator';

export class UserProjectSkillUpdateDto {
  @ApiProperty({ example: 'Description', required: false })
  @Expose()
  @IsOptional()
  description?: string;
}