# swagger-spec.json
{"openapi":"3.0.0","paths":{"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}},"tags":["Authentication"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmptyResponseDto"}}}}},"tags":["Authentication"],"security":[{"refreshToken":[]}]}},"/auth/refresh":{"post":{"operationId":"AuthController_refreshTokens","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}},"tags":["Authentication"],"security":[{"refreshToken":[]}]}},"/users/me":{"get":{"operationId":"UserController_findCurrent","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/users/{username}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"username","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}},"tags":["Users"]}},"/userProfile/general":{"put":{"operationId":"UserProfileController_updateOne","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfileUpdateDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}},"tags":["UserProfile"],"security":[{"bearer":[]}]}},"/userProfile/profileImage":{"put":{"operationId":"UserProfileController_uploadFile","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":""}},"tags":["UserProfile"],"security":[{"bearer":[]}]}},"/projects/{id}":{"put":{"operationId":"ProjectController_updateOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectUpdateDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectResponseDto"}}}}},"tags":["Projects"],"security":[{"bearer":[]}]}},"/projects/{id}/image":{"put":{"operationId":"ProjectController_updateProjectImage","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":""}},"tags":["Projects"],"security":[{"bearer":[]}]}},"/userProject":{"post":{"operationId":"UserProjectController_addOne","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectCreateDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmptyResponseDto"}}}},"404":{"description":""}},"tags":["User - Project"],"security":[{"bearer":[]}]}},"/userProject/{id}":{"put":{"operationId":"UserProjectController_updateOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProjectUpdateDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProjectUpdateDto"}}}}},"tags":["User - Project"],"security":[{"bearer":[]}]},"delete":{"operationId":"UserProjectController_deleteOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"404":{"description":""},"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmptyResponseDto"}}}}},"tags":["User - Project"],"security":[{"bearer":[]}]}},"/skills":{"get":{"operationId":"SkillController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SkillResponseDto"}}}}}},"tags":["Skills"]}},"/skills/{id}":{"get":{"operationId":"SkillController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SkillResponseDto"}}}},"404":{"description":""}},"tags":["Skills"]}},"/userSkill/{skillId}":{"post":{"operationId":"UserSkillController_addOne","parameters":[{"name":"skillId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSkillResponseDto"}}}},"404":{"description":""}},"tags":["User - Skill"],"security":[{"bearer":[]}]}},"/userSkill/{id}":{"put":{"operationId":"UserSkillController_updateOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSkillUpdateDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSkillResponseDto"}}}}},"tags":["User - Skill"],"security":[{"bearer":[]}]},"delete":{"operationId":"UserSkillController_deleteOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"204":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmptyResponseDto"}}}}},"tags":["User - Skill"],"security":[{"bearer":[]}]}},"/userProjectSkill":{"post":{"operationId":"UserProjectSkillController_addOne","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProjectSkillCreateDto"}}}},"responses":{"201":{"description":""}},"tags":["UserProject - Skill"],"security":[{"bearer":[]}]}},"/userProjectSkill/{id}":{"put":{"operationId":"UserProjectSkillController_updateOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProjectSkillUpdateDto"}}}},"responses":{"200":{"description":""}},"tags":["UserProject - Skill"],"security":[{"bearer":[]}]},"delete":{"operationId":"UserProjectSkillController_deleteOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"204":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmptyResponseDto"}}}},"404":{"description":""}},"tags":["UserProject - Skill"],"security":[{"bearer":[]}]}}},"info":{"title":"Levelup API","description":"This is a swagger documentation for levelup API","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginDto":{"type":"object","properties":{"username":{"type":"string","example":"noffirl"},"password":{"type":"string","example":"test22848"}},"required":["username","password"]},"AuthResponseDto":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}},"required":["accessToken","refreshToken"]},"RegisterDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"EmptyResponseDto":{"type":"object","properties":{"status":{"type":"number","example":204}},"required":["status"]},"SkillResponseDto":{"type":"object","properties":{"id":{"type":"number","example":1},"title":{"type":"string","example":"JavaScript"}},"required":["id","title"]},"ProjectSkillResponseDto":{"type":"object","properties":{"id":{"type":"number","example":1},"skill":{"$ref":"#/components/schemas/SkillResponseDto"},"description":{"type":"string","example":"description"}},"required":["id","skill","description"]},"ProjectResponseDto":{"type":"object","properties":{"id":{"type":"number","example":1},"title":{"type":"string","example":"Project title"},"description":{"type":"string","example":"Project description"},"projectSkills":{"type":"array","items":{"$ref":"#/components/schemas/ProjectSkillResponseDto"}},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"image":{"type":"string"}},"required":["id","title","description","projectSkills","createdAt","updatedAt","image"]},"UserProjectSkillResponseDto":{"type":"object","properties":{"id":{"type":"number","example":1},"skill":{"$ref":"#/components/schemas/SkillResponseDto"},"description":{"type":"string","example":"Brief description"}},"required":["id","skill","description"]},"UserProjectResponseDto":{"type":"object","properties":{"id":{"type":"number","example":1},"project":{"$ref":"#/components/schemas/ProjectResponseDto"},"role":{"type":"string","example":"Front-End"},"description":{"type":"string","example":"Brief description"},"skills":{"type":"array","items":{"$ref":"#/components/schemas/UserProjectSkillResponseDto"}}},"required":["id","project","role","description","skills"]},"UserSkillResponseDto":{"type":"object","properties":{"id":{"type":"number","example":1},"skill":{"$ref":"#/components/schemas/SkillResponseDto"},"learnedAt":{"format":"date-time","type":"string"},"description":{"type":"string","example":"Brief description"}},"required":["id","skill","learnedAt","description"]},"UserResponseDto":{"type":"object","properties":{"id":{"type":"number","example":1},"username":{"type":"string","example":"noffirl"},"userProjects":{"type":"array","items":{"$ref":"#/components/schemas/UserProjectResponseDto"}},"userSkills":{"type":"array","items":{"$ref":"#/components/schemas/UserSkillResponseDto"}},"firstName":{"type":"string","example":"Ihor"},"lastName":{"type":"string","example":"Didunik"},"email":{"type":"string","example":"diduniki@gmail.com"},"profileImage":{"type":"string","example":"undefined/image.jpg"},"bio":{"type":"string","example":"bio in html format"},"phoneNumber":{"type":"string","example":"+353-(085)-194-55-49"},"address":{"type":"string","example":"Bulgaria, Varna"},"dateOfBirth":{"format":"date-time","type":"string"}},"required":["id","username","userProjects","userSkills","firstName","lastName","email","profileImage","bio","phoneNumber","address","dateOfBirth"]},"UserProfileUpdateDto":{"type":"object","properties":{"firstName":{"type":"string","example":"Ihor"},"lastName":{"type":"string","example":"Didunik"},"email":{"type":"string","example":"test.mail.com"},"bio":{"type":"string","example":"html bio"},"phoneNumber":{"type":"string"},"address":{"type":"string"},"dateOfBirth":{"format":"date-time","type":"string"}}},"ProjectUpdateDto":{"type":"object","properties":{"title":{"type":"string","example":"Crypto Wallet"},"description":{"type":"string","example":"Amazing de-fi application bla bla bla"},"shortDescription":{"type":"string","example":"Amazing de-fi application..."}}},"ProjectCreateDto":{"type":"object","properties":{"title":{"type":"string","example":"Crypto Wallet"},"description":{"type":"string","example":"Amazing de-fi application bla bla bla"},"shortDescription":{"type":"string","example":"Amazing de-fi application..."}},"required":["title"]},"UserProjectUpdateDto":{"type":"object","properties":{"role":{"type":"string","example":"Back-End"},"description":{"type":"string","example":"Brief description"},"shortDescription":{"type":"string","example":"Brief description"}}},"UserSkillUpdateDto":{"type":"object","properties":{"description":{"type":"string","example":"Description"},"learnedAt":{"format":"date-time","type":"string"}}},"UserProjectSkillCreateDto":{"type":"object","properties":{"userProjectId":{"type":"number","example":1},"skillId":{"type":"number","example":1}},"required":["userProjectId","skillId"]},"UserProjectSkillUpdateDto":{"type":"object","properties":{"description":{"type":"string","example":"Description"}}}}}}